app/requirements.txt
fastapi
uvicorn
passlib
pymysql
sqlalchemy
python-multipart
pyjwt
tensorflow
numpy
pandas
tensorflow_hub
